[TESTS] Since the *.typo3.org, we can resolve everything
We cannot test if we can resolve the host name `otrs.vagrant`,
because thanks to the new `*.typo3.org` entry, every hostname
points to the proxies, if we have a search domain `typo3.org`
configured (which we have on the CI server).

Therefore, always update `/etc/hosts` to add `otrs.vagrant`.